Refactor Footer component and create Footer1 for detailed charity information
